requirements.txt
google-genai
watchdog
python-dotenv
